move all base component js into abstract folder
