+++ src/components/m-footer-social/index.js
@@ -1,6 +1,6 @@
 import classnames from 'classnames';
 import getAttribute from '../../js/get-attribute';
-import BaseComponentGlobal from '../../js/base-component-global';
+import BaseComponentGlobal from '../../js/abstract/base-component-global';
 import styles from './index.scss';
 import template from './_template';
 import wcdomready from '../../js/wcdomready';
