+++ src/components/m-header-main/index.js
@@ -1,5 +1,5 @@
 import classnames from 'classnames';
-import BaseComponentGlobal from '../../js/base-component-global';
+import BaseComponentGlobal from '../../js/abstract/base-component-global';
 // import the styles used for this component
 import styles from './index.scss';
 // import the template used for this component
