+++ src/components/m-header-meta-right/index.js
@@ -1,4 +1,4 @@
-import BaseComponentGlobal from '../../js/base-component-global';
+import BaseComponentGlobal from '../../js/abstract/base-component-global';
 // import the styles used for this component
 import styles from './index.scss';
 // import the template used for this component
