+++ src/components/m-header-sub-navigation/index.js
@@ -1,7 +1,7 @@
 import classnames from 'classnames';
 import styles from './index.scss';
 import template from './_template';
-import BaseComponentGlobal from '../../js/base-component-global';
+import BaseComponentGlobal from '../../js/abstract/base-component-global';
 import wcdomready from '../../js/wcdomready';
 
 class AXASubNavigation extends BaseComponentGlobal {
