+++ src/components/m-link/index.js
@@ -1,6 +1,6 @@
 import styles from './index.scss';
 import template from './_template';
-import BaseComponentGlobal from '../../js/base-component-global';
+import BaseComponentGlobal from '../../js/abstract/base-component-global';
 import wcdomready from '../../js/wcdomready';
 
 class AXALink extends BaseComponentGlobal {
