+++ src/components/o-header/index.js
@@ -1,5 +1,5 @@
 import styles from './index.scss';
-import BaseComponentGlobal from '../../js/base-component-global';
+import BaseComponentGlobal from '../../js/abstract/base-component-global';
 import wcdomready from '../../js/wcdomready';
 
 class AXAHeader extends BaseComponentGlobal {
