+++ src/js/abstract/base-component-global.js
