+++ src/js/abstract/base-component-shadow.js
