+++ src/js/abstract/component-types.js
