+++ stack/tasks/create-a-m-o.js
@@ -70,7 +70,7 @@ const writeIndexJs = (path, _name) => {
 
   fs.writeFileSync(
     `${path}/index.js`,
-    outdent`import { BaseComponentGlobal } from '../../js/base-component-global';
+    outdent`import { BaseComponentGlobal } from '../../js/abstract/base-component-global';
       // import the styles used for this component
       import styles from './index.scss';
       // import the template used for this component
